organizing and preparing for documentation

+++ towers/uhd_scripts/fft_file_analizer.py
@@ -21,9 +21,12 @@ class Analizer:
 		self.fft_size = None
 		self.gain = None
 		self.row_map = None
+		self.remove = None
 		self.filename = filename
 		self.row_mapper()
 		self.freq = sorted(np.fft.fftfreq(self.fft_size, d = 1.0/self.samp_rate))
+	'''
+	'''
 	def decompose(self):
 		parts = self.filename.split('_')
 		samp_rate = parts[0].strip()
@@ -34,6 +37,9 @@ class Analizer:
 		fileparts = FileParts(samp_rate,stamp,frequency,extension)
 		self.fileparts = fileparts
 		return fileparts
+	'''
+	converts a row of meta file data to a sample object
+	'''
 	def row_converter(self,row):
 		if self.samp_rate is None:
 			self.samp_rate = row.samp_rate
@@ -41,15 +47,26 @@ class Analizer:
 			self.fft_size = row.fft_size
 		if self.gain is None:
 			self.gain = row.gain
+		if self.remove is None:
+			self.remove = row.remove
 		return Sample(row.time,row.c_freq,row.filename)
+	'''
+	converts meta file to series of Sample object 
+	'''
 	def row_mapper(self):
 		data = pd.read_csv(self.filename).apply(self.row_converter,1)
 		os.remove(self.filename)
 		self.filename= None
 		self.row_map = data
 		return data
+	'''
+	takes in complex numpy array and multiples complex conjugates
+	'''
 	def multiply_conj(self,cmplx):
 		return abs(cmplx*np.conjugate(cmplx))
+	'''
+	Averages all the fft samples in a sample data
+	'''
 	def fft_average(self,sample):
 		filename = sample.filename
 		vect = np.vectorize(self.multiply_conj)
@@ -59,18 +76,27 @@ class Analizer:
 		segments = data.reshape(number_of_ffts,self.fft_size)
 		means = np.average(segments,axis=0)
 		return means
+	'''
+	Maps frequencies to samples, returning a array of shape (2,fft-size)
+	'''
 	def freq_map(self,sample):
 		freq = np.add(self.freq, sample.center_freq)
 		segment = self.fft_average(sample)
 		return np.vstack((segment.T,freq.T)).T
+	'''
+	Averages the ffts and saves them to a file
+	'''
 	def pw_frq_pair_to_csv(self, sample):
 		f = str(sample.time_stamp)+'_'+str(int(sample.center_freq))+'.csv'
 		pd.DataFrame(self.freq_map(sample)).to_csv('collected_data/'+f)
-		os.remove(sample.filename)
-
-
+		if self.remove:
+			os.remove(sample.filename)
+	'''
+	'''
 	def analyze(self,filename):
 		return None
+	'''
+	'''
 	def plot_average(self,sample_index = 0):
 		sample = self.row_map[sample_index]
 		means = self.fft_average(sample)
